requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/dev.txt requirements/ci.txt requirements/dev.in
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.3            # via black
attrs==19.3.0             # via -r requirements/ci.txt, jsonschema
babel==2.8.0              # via sphinx
beautifulsoup4==4.9.0     # via -r requirements/ci.txt, webtest
black==20.8b1             # via -r requirements/dev.in
bump2version==1.0.1       # via bumpversion
bumpversion==0.6.0        # via -r requirements/dev.in
certifi==2020.4.5.1       # via -r requirements/ci.txt, elastic-apm, requests, sentry-sdk
chardet==3.0.4            # via -r requirements/ci.txt, requests
click==7.1.2              # via black, pip-tools
commonmark==0.9.1         # via recommonmark
coreapi==2.3.3            # via -r requirements/ci.txt, drf-yasg
coreschema==0.0.4         # via -r requirements/ci.txt, coreapi, drf-yasg
coverage==4.5.4           # via -r requirements/ci.txt
cssselect==1.1.0          # via -r requirements/ci.txt, pyquery
django-admin-index==1.4.0  # via -r requirements/ci.txt
django-appconf==1.0.4     # via -r requirements/ci.txt, django-axes
django-axes==5.3.1        # via -r requirements/ci.txt
django-choices==1.7.1     # via -r requirements/ci.txt, vng-api-common, zgw-consumers
django-compat==1.0.15     # via -r requirements/ci.txt, django-hijack, django-hijack-admin
django-debug-toolbar==2.2  # via -r requirements/dev.in
django-extensions==2.2.9  # via -r requirements/dev.in
django-filter==2.4.0      # via -r requirements/ci.txt, vng-api-common
django-hijack-admin==2.1.10  # via -r requirements/ci.txt
django-hijack==2.1.10     # via -r requirements/ci.txt, django-hijack-admin
django-ipware==2.1.0      # via -r requirements/ci.txt, django-axes
django-jenkins==0.110.0   # via -r requirements/ci.txt
django-ordered-model==3.3.0  # via -r requirements/ci.txt, django-admin-index
django-redis==4.11.0      # via -r requirements/ci.txt
django-relativedelta==1.1.2  # via -r requirements/ci.txt, zgw-consumers
django-rest-framework-condition==0.1.1  # via -r requirements/ci.txt, vng-api-common
django-rosetta==0.9.4     # via -r requirements/ci.txt
django-sniplates==0.7.0   # via -r requirements/ci.txt
django-solo==1.1.3        # via -r requirements/ci.txt, vng-api-common, zgw-consumers
django-webtest==1.9.7     # via -r requirements/ci.txt
django==2.2.12            # via -r requirements/ci.txt, django-appconf, django-axes, django-choices, django-debug-toolbar, django-filter, django-jenkins, django-redis, django-relativedelta, django-rest-framework-condition, django-rosetta, django-sniplates, drf-nested-routers, drf-yasg, vng-api-common, zgw-consumers
djangorestframework-camel-case==1.2.0  # via -r requirements/ci.txt, vng-api-common
djangorestframework-gis==0.16  # via -r requirements/ci.txt
djangorestframework==3.9.4  # via -r requirements/ci.txt, djangorestframework-gis, drf-nested-routers, drf-yasg, vng-api-common
docutils==0.16            # via recommonmark, sphinx
drf-nested-routers==0.91  # via -r requirements/ci.txt, vng-api-common
drf-yasg==1.16.0          # via -r requirements/ci.txt, vng-api-common
elastic-apm==5.5.2        # via -r requirements/ci.txt
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via -r requirements/ci.txt
faker==4.0.3              # via -r requirements/ci.txt, factory-boy
flake8==3.7.9             # via -r requirements/dev.in
freezegun==0.3.15         # via -r requirements/ci.txt
gemma-zds-client==0.13.3  # via -r requirements/ci.txt, vng-api-common, zgw-consumers
idna==2.9                 # via -r requirements/ci.txt, requests
imagesize==1.2.0          # via sphinx
inflection==0.4.0         # via -r requirements/ci.txt, drf-yasg
iso-639==0.4.5            # via -r requirements/ci.txt, vng-api-common
isodate==0.6.0            # via -r requirements/ci.txt, vng-api-common
isort==5.5.3              # via -r requirements/dev.in
itypes==1.2.0             # via -r requirements/ci.txt, coreapi
jinja2==2.11.2            # via -r requirements/ci.txt, coreschema, sphinx
jsonschema==3.2.0         # via -r requirements/ci.txt
lxml==4.5.0               # via -r requirements/ci.txt, pyquery
markupsafe==1.1.1         # via -r requirements/ci.txt, jinja2
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
oyaml==1.0                # via -r requirements/ci.txt, vng-api-common
packaging==20.3           # via sphinx
pathspec==0.8.0           # via black
pillow==7.1.1             # via -r requirements/ci.txt
pip-tools==5.4.0          # via -r requirements/dev.in
polib==1.1.0              # via -r requirements/ci.txt, django-rosetta
psycopg2==2.8.5           # via -r requirements/ci.txt
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via sphinx, sphinx-tabs
pyjwt==1.7.1              # via -r requirements/ci.txt, gemma-zds-client, vng-api-common
pyparsing==2.4.7          # via packaging
pyquery==1.4.1            # via -r requirements/ci.txt
pyrsistent==0.16.0        # via -r requirements/ci.txt, jsonschema
python-dateutil==2.8.1    # via -r requirements/ci.txt, django-relativedelta, faker, freezegun
python-decouple==3.3      # via -r requirements/ci.txt
python-dotenv==0.12.0     # via -r requirements/ci.txt
pytz==2019.3              # via -r requirements/ci.txt, babel, django
pyyaml==5.3.1             # via -r requirements/ci.txt, gemma-zds-client, oyaml, vng-api-common
recommonmark==0.6.0       # via -r requirements/dev.in
redis==3.4.1              # via -r requirements/ci.txt, django-redis
regex==2020.4.4           # via black
requests-mock==1.7.0      # via -r requirements/ci.txt
requests==2.23.0          # via -r requirements/ci.txt, coreapi, django-rosetta, gemma-zds-client, requests-mock, sphinx, vng-api-common, zgw-consumers
ruamel.yaml.clib==0.2.0   # via -r requirements/ci.txt, ruamel.yaml
ruamel.yaml==0.16.10      # via -r requirements/ci.txt, drf-yasg
sentry-sdk==0.14.3        # via -r requirements/ci.txt
six==1.14.0               # via -r requirements/ci.txt, django-choices, django-compat, django-extensions, django-rosetta, drf-yasg, freezegun, isodate, jsonschema, packaging, pip-tools, pyrsistent, python-dateutil, requests-mock, webtest
snowballstemmer==2.0.0    # via sphinx
soupsieve==2.0            # via -r requirements/ci.txt, beautifulsoup4
sphinx-rtd-theme==0.4.3   # via -r requirements/dev.in
sphinx-tabs==1.3.0        # via -r requirements/dev.in
sphinx==3.0.1             # via -r requirements/dev.in, recommonmark, sphinx-rtd-theme, sphinx-tabs
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.3.1           # via -r requirements/ci.txt, django, django-debug-toolbar
tblib==1.6.0              # via -r requirements/ci.txt
text-unidecode==1.3       # via -r requirements/ci.txt, faker
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
unidecode==1.1.1          # via -r requirements/ci.txt, vng-api-common
uritemplate==3.0.1        # via -r requirements/ci.txt, coreapi, drf-yasg
urllib3==1.25.8           # via -r requirements/ci.txt, elastic-apm, requests, sentry-sdk
uwsgi==2.0.18             # via -r requirements/ci.txt
vng-api-common==1.5.6     # via -r requirements/ci.txt
waitress==1.4.3           # via -r requirements/ci.txt, webtest
webob==1.8.6              # via -r requirements/ci.txt, webtest
webtest==2.0.34           # via -r requirements/ci.txt, django-webtest
zgw-consumers==0.13.1     # via -r requirements/ci.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
